fixed styles and rename component
