+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Projects\\Test-tasks\\user-list\\src\\App.js":"1","D:\\Projects\\Test-tasks\\user-list\\src\\components\\FilteredUsers\\FilteredUsers.js":"2","D:\\Projects\\Test-tasks\\user-list\\src\\components\\NewUser\\NewUser.js":"3","D:\\Projects\\Test-tasks\\user-list\\src\\components\\UserCards\\UserCards.js":"4","D:\\Projects\\Test-tasks\\user-list\\src\\api.js":"5","D:\\Projects\\Test-tasks\\user-list\\src\\index.js":"6"},{"size":1572,"mtime":1606678194064,"results":"7","hashOfConfig":"8"},{"size":315,"mtime":1606656278378,"results":"9","hashOfConfig":"8"},{"size":437,"mtime":1606506606085,"results":"10","hashOfConfig":"8"},{"size":1014,"mtime":1606723215421,"results":"11","hashOfConfig":"8"},{"size":146,"mtime":1606465696328,"results":"12","hashOfConfig":"8"},{"size":157,"mtime":1606410049267,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"osceki",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Projects\\Test-tasks\\user-list\\src\\App.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\components\\FilteredUsers\\FilteredUsers.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\components\\NewUser\\NewUser.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\components\\UserCards\\UserCards.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\api.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\index.js",[]]
+[{"D:\\Projects\\Test-tasks\\user-list\\src\\api.js":"1","D:\\Projects\\Test-tasks\\user-list\\src\\index.js":"2","D:\\Projects\\Test-tasks\\user-list\\src\\components\\NewUser\\NewUser.jsx":"3","D:\\Projects\\Test-tasks\\user-list\\src\\components\\UserCards\\UserCards.jsx":"4","D:\\Projects\\Test-tasks\\user-list\\src\\App.jsx":"5","D:\\Projects\\Test-tasks\\user-list\\src\\components\\SearchOfUsers\\SearchOfUsers.jsx":"6"},{"size":146,"mtime":1606465696328,"results":"7","hashOfConfig":"8"},{"size":162,"mtime":1606725496475,"results":"9","hashOfConfig":"8"},{"size":437,"mtime":1606725337150,"results":"10","hashOfConfig":"8"},{"size":1318,"mtime":1606725903300,"results":"11","hashOfConfig":"8"},{"size":1584,"mtime":1606725999897,"results":"12","hashOfConfig":"8"},{"size":315,"mtime":1606725960849,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"osceki",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Projects\\Test-tasks\\user-list\\src\\api.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\index.js",[],"D:\\Projects\\Test-tasks\\user-list\\src\\components\\NewUser\\NewUser.jsx",[],"D:\\Projects\\Test-tasks\\user-list\\src\\components\\UserCards\\UserCards.jsx",[],"D:\\Projects\\Test-tasks\\user-list\\src\\App.jsx",[],"D:\\Projects\\Test-tasks\\user-list\\src\\components\\SearchOfUsers\\SearchOfUsers.jsx",[]]
