+++ debug.log
@@ -14,3 +14,5 @@
 [1129/210716.763:ERROR:crash_report_database_win.cc(457)] unexpected header
 [1129/213445.356:ERROR:crash_report_database_win.cc(457)] unexpected header
 [1130/100024.845:ERROR:crash_report_database_win.cc(457)] unexpected header
+[1130/103011.848:ERROR:crash_report_database_win.cc(457)] unexpected header
+[1130/104728.622:ERROR:crash_report_database_win.cc(457)] unexpected header
