+++ public/index.html
@@ -15,7 +15,7 @@
     >
     <title>React App</title>
   </head>
-  <body>
+  <body style="background-color:#1d1d1d;">
     <div id="root"></div>
   </body>
 </html>
