+++ src/App.css
@@ -3,7 +3,6 @@
   flex-direction: column;
   align-items: center;
   flex-wrap: wrap;
-  background-color: #1d1d1d;
   color: #fff;
 }
