+++ src/api.js
@@ -1,5 +1,5 @@
 export async function getUsers(count) {
-  const respons = await fetch(`https://randomuser.me/api/?results=${count}`);
+  const response = await fetch(`https://randomuser.me/api/?results=${count}`);
 
-  return respons.json();
+  return response.json();
 }
