+++ src/components/UserCards/UserCards.css
@@ -40,3 +40,8 @@
   margin: 0 auto;
   padding: 5px;
 }
+
+.user__alert {
+  padding: 20px;
+  font-size: 20px;
+}
